Controller refactor
